added arg to plot histogram while selecting model
