+++ src/modeling/model_selector_helper.py
@@ -5,6 +5,8 @@ import numpy as np
 
 from dotenv import load_dotenv
 from model import Model
+import matplotlib.pyplot as plt
+import seaborn as sns
 
 load_dotenv()
 root = os.getenv("root")
@@ -70,3 +72,48 @@ def select_models(keys, clustering, n):
         print(f"Selected {best_model}")
         selection.append(best_model)
     return selection
+
+
+def unpack_policy_group_dir(folder):
+    n = int(folder[: folder.index("_")])
+    return n
+
+
+def get_viable_models(n: int, coverage_filter: float):
+    dir = f"data/mappers/{n}_policy_groups/"
+    dir = os.path.join(root, dir)
+    files = os.listdir(dir)
+    models = []
+    for file in files:
+        file = os.path.join(dir, file)
+        model = Model(file)
+        N = len(model.tupper.clean)
+        num_unclustered_items = len(model.unclustered_items)
+        if num_unclustered_items / N <= 1 - coverage_filter:
+            models.append(file)
+    print(
+        f"{n}_policy_groups: {np.round(len(models)/len(files)*100,1)}% of models had at least {coverage_filter*100}% coverage."
+    )
+    return models
+
+
+def plot_mapper_histogram(coverage_filter=0.8):
+    mappers = os.path.join(root, "data/mappers/")
+    policy_groups = os.listdir(mappers)  # get a list of folder names in the directory
+    counts = {}  # initialize an empty list to store the number of files in each folder
+    for folder in policy_groups:
+        n = unpack_policy_group_dir(folder)
+        models = get_viable_models(n, coverage_filter)
+        counts[n] = len(models)
+    keys = list(counts.keys())
+    keys.sort()
+    sorted_counts = {i: counts[i] for i in keys}
+    # plot the histogram
+    fig = plt.figure(figsize=(15, 10))
+    ax = sns.barplot(
+        x=list(sorted_counts.keys()),
+        y=list(sorted_counts.values()),
+    )
+    ax.set(xlabel="Number of Policy Groups", ylabel="Number of Viable Models")
+    plt.show()
+    return fig
